+++ SS-Microservice.Services.Basket/Domain/Entities/Basket.cs
@@ -2,7 +2,7 @@
 
 namespace SS_Microservice.Services.Basket.Domain.Entities
 {
-    public class Basket : BaseAuditableEntity<int>
+    public class Basket : BaseAuditableEntity<long>
     {
         public string UserId { get; set; }
